+++ leia-common/src/test/java/com/grookage/leia/common/builder/SchemaBuilderTest.java
@@ -1,23 +1,12 @@
 package com.grookage.leia.common.builder;
 
 import com.grookage.leia.common.LeiaTestUtils;
-import com.grookage.leia.common.stubs.NestedStub;
-import com.grookage.leia.common.stubs.RecordStub;
-import com.grookage.leia.common.stubs.TestEnum;
-import com.grookage.leia.common.stubs.TestObjectStub;
-import com.grookage.leia.common.stubs.TestParameterizedStub;
-import com.grookage.leia.common.stubs.TestRawCollectionStub;
+import com.grookage.leia.common.stubs.*;
 import com.grookage.leia.models.annotations.SchemaDefinition;
 import com.grookage.leia.models.annotations.attribute.Optional;
 import com.grookage.leia.models.annotations.attribute.qualifiers.Encrypted;
 import com.grookage.leia.models.annotations.attribute.qualifiers.PII;
-import com.grookage.leia.models.attributes.ArrayAttribute;
-import com.grookage.leia.models.attributes.EnumAttribute;
-import com.grookage.leia.models.attributes.IntegerAttribute;
-import com.grookage.leia.models.attributes.MapAttribute;
-import com.grookage.leia.models.attributes.ObjectAttribute;
-import com.grookage.leia.models.attributes.SchemaAttribute;
-import com.grookage.leia.models.attributes.StringAttribute;
+import com.grookage.leia.models.attributes.*;
 import com.grookage.leia.models.qualifiers.EncryptedQualifier;
 import com.grookage.leia.models.qualifiers.PIIQualifier;
 import com.grookage.leia.models.schema.SchemaType;