+++ leia-models/src/main/java/com/grookage/leia/models/request/ValidateSchemaRequest.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
 @Data
@@ -16,6 +17,7 @@ import javax.validation.constraints.NotNull;
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ValidateSchemaRequest {
+    @Valid
     @NotNull
     private SchemaKey schemaKey;
     @NotNull